requirements.txt
streamlit
ollama
torch
opencv-python
Pillow
plotly
fastapi
uvicorn